Update delete-queue-rule.php
update log: 112724

+++ delete-queue-rule.php
@@ -1,24 +1,24 @@
 <?php 
 
-// session_start();
+session_start();
 
 // Contents down below will only function if the user is set (Administrator)
-// if (isset($user) && ($user["role_fld"]) == "Administrator") {
+if (isset($user) && ($user["role_fld"]) == "Administrator") {
 
     // Database connection
     $mysqli = require __DIR__ . "/database.php";
 
-    // Check if the request is via POST
+    // Checks if the request is via POST
     if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-        // Retrieve the posted username from the AJAX request
+        // Retrieves the posted username from the AJAX request
         $username = trim($_POST['username']);
 
-        // Prepare the SQL query
+        // Prepares the SQL query
         $stmt = $mysqli->prepare("DELETE FROM queue_rules_tbl WHERE queue_id_fld = ?");
         $stmt->bind_param("s", $username);
         $stmt->execute();
         
-        // Return the result as a JSON response
+        // Returns the result as a JSON response
         echo json_encode(['success' => true, 'message' => 'Deleted user successfully']);
 
         $stmt->close();
@@ -29,9 +29,10 @@
 
 
 
-// }else{
+}else{
     
-//     header("Location: index.php");
-//     exit;
+    header("Location: index.php");
+    exit;
             
-//         } ?>
+        } 
+?>
